moving tests from board_test.rb to neighboring_cells_test and adjacent_empties_test
